updated to use more modern dialogs
