+++ javascript_nodejs/10.prompt-validations/package.json
@@ -11,9 +11,9 @@
   "author": "Microsoft",
   "license": "MIT",
   "dependencies": {
-    "botbuilder": "^4.0.0-preview1.2",
-    "botbuilder-dialogs": "^4.0.0-preview1.2",
-    "botframework-config": "^4.0.0-preview1.3",
+    "botbuilder": "^4.0.0-preview1.1140",
+    "botbuilder-dialogs": "^4.0.0-preview1.1140",
+    "botframework-config": "^4.0.0-preview1.1140",
     "dotenv": "^6.0.0",
     "path": "^0.12.7",
     "restify": "^7.2.1"
